refactor : 의존성에따른 패키지분리
