+++ src/main/java/blackjack/BlackjackApplication.java
@@ -2,12 +2,11 @@ package blackjack;
 
 import java.util.Scanner;
 
-import blackjack.player.card.CardFactory;
+import blackjack.card.CardFactory;
 import blackjack.view.InputView;
-import blackjack.view.OutputView;
 
 public class BlackjackApplication {
 	public static void main(String[] args) {
-		new BlackjackGame(CardFactory.getInstance(), new InputView(new Scanner(System.in)), new OutputView()).run();
+		new BlackjackController(CardFactory.getInstance(), new InputView(new Scanner(System.in))).run();
 	}
 }
