+++ src/main/java/blackjack/card/GameResult.java
@@ -1,4 +1,4 @@
-package blackjack.player.card;
+package blackjack.card;
 
 import java.util.Arrays;
