+++ src/main/java/blackjack/card/component/CardNumber.java
@@ -1,4 +1,4 @@
-package blackjack.player.card.component;
+package blackjack.card.component;
 
 public enum CardNumber {
 	ACE(1, "A"),
