--- src/main/java/blackjack/player/card/component/Symbol.java
@@ -1,18 +1,0 @@
-package blackjack.player.card.component;
-
-public enum Symbol {
-    HEART("하트"),
-    SPADE("스페이드"),
-    DIAMOND("다이아몬드"),
-    CLUB("클럽");
-
-    private final String name;
-
-    Symbol(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
