--- src/main/java/blackjack/view/DrawRequest.java
@@ -1,36 +1,0 @@
-package blackjack.view;
-
-import java.util.Arrays;
-
-public class DrawRequest {
-    private final boolean result;
-
-    public DrawRequest(String command) {
-        this.result = PermittedRequest.findByCommand(command)
-                .result;
-    }
-
-    public boolean isDraw() {
-        return result;
-    }
-
-    private enum PermittedRequest {
-        Y("y", true),
-        N("n", false);
-
-        private final String command;
-        private final boolean result;
-
-        PermittedRequest(String command, boolean result) {
-            this.command = command;
-            this.result = result;
-        }
-
-        public static PermittedRequest findByCommand(String command) {
-            return Arrays.stream(values())
-                    .filter(permittedRequest -> permittedRequest.command.equals(command))
-                    .findFirst()
-                    .orElseThrow(() -> new IllegalArgumentException(String.format("%s는 올바른 요청이 아닙니다.", command)));
-        }
-    }
-}
