+++ src/test/java/blackjack/card/CardFactoryTest.java
@@ -1,0 +1,10 @@
+package blackjack.card;
+
+import org.junit.jupiter.api.Test;
+
+class CardFactoryTest {
+
+	@Test
+	void drawCard() {
+	}
+}
