+++ src/test/java/blackjack/player/GamblerTest.java
@@ -1,20 +1,21 @@
 package blackjack.player;
 
-import blackjack.player.card.CardBundle;
+import static org.assertj.core.api.Assertions.*;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import blackjack.card.CardBundle;
 
 class GamblerTest {
 
-    @DisplayName("겜블러는 게임 결과를 만들어낼수 없습니다.")
-    @Test
-    void getReport() {
-        Player gambler1 = new Gambler(new CardBundle(), "bebop");
-        Player gambler2 = new Gambler(new CardBundle(), "allen");
+	@DisplayName("겜블러는 게임 결과를 만들어낼수 없습니다.")
+	@Test
+	void getReport() {
+		Player gambler1 = new Gambler(new CardBundle(), "bebop");
+		Player gambler2 = new Gambler(new CardBundle(), "allen");
 
-        assertThatThrownBy(() -> gambler1.getReport(gambler2))
-                .isInstanceOf(UnsupportedOperationException.class);
-    }
+		assertThatThrownBy(() -> gambler1.getReport(gambler2))
+			.isInstanceOf(UnsupportedOperationException.class);
+	}
 }
