+++ src/test/java/blackjack/player/PlayerTest.java
@@ -1,19 +1,20 @@
 package blackjack.player;
 
-import blackjack.player.card.Card;
-import blackjack.player.card.CardBundle;
-import blackjack.player.card.component.CardNumber;
-import blackjack.player.card.component.Symbol;
+import static blackjack.card.CardBundleHelper.*;
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.stream.Stream;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.stream.Stream;
-
-import static blackjack.player.card.CardBundleHelper.aCardBundle;
-import static org.assertj.core.api.Assertions.assertThat;
+import blackjack.card.Card;
+import blackjack.card.CardBundle;
+import blackjack.card.component.CardNumber;
+import blackjack.card.component.Symbol;
 
 class PlayerTest {
